% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aggregate-metric.R
\name{aggregateMetric}
\alias{aggregateMetric}
\title{Aggregates data from Facebook page by period of time}
\usage{
aggregateMetric(page, metric, FUN)
}
\arguments{
\item{page}{Data frame with Facebook page data}

\item{metric}{Either "likes", "comments" or "shares"}

\item{FUN}{function used to aggregate. Generally, "sum"}
}
\description{
Aggregates data from Facebook page by period of time
}
\examples{
\dontrun{
## example of Facebook data
  page = getPage("barackobama", token=my_oauth, n=300)
	 page$datetime = formatFBdate(page$created_time)
  results = aggregate.metric(page, metric="likes")
}
}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}
}

