% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/count-tweet-months.R
\name{countMonthsTweets}
\alias{countMonthsTweets}
\title{Count the number of tweets sent by month}
\usage{
countMonthsTweets(dates)
}
\arguments{
\item{dates}{Vector of dates in Twitter format}
}
\description{
Count the number of tweets sent by month
}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}
}

