% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/count-post-weeks.R, R/count-tweet-weeks.R
\name{countWeeksPosts}
\alias{countWeeksPosts}
\alias{countWeeksTweets}
\title{Count the number of tweets sent by week}
\usage{
countWeeksPosts(dates)

countWeeksTweets(dates)
}
\arguments{
\item{dates}{Vector of dates in Twitter format}

\item{dates}{Vector of dates in Twitter format}
}
\description{
Count the number of tweets sent by week

Count the number of tweets sent by week
}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}

Pablo Barbera \email{pablo.barbera@nyu.edu}
}

