% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-common-hashtags.R
\name{getCommonHashtags}
\alias{getCommonHashtags}
\title{Find the N most common hashtags in a string vector}
\usage{
getCommonHashtags(text, n = 20)
}
\arguments{
\item{text}{Vector containing tweet text}

\item{n}{number of hashtags to display}
}
\description{
Find the N most common hashtags in a string vector
}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}
}

