% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-tweets.R
\name{parseTweets}
\alias{parseTweets}
\title{Modified version of parseTweets that will return retweet
count for tweet (not retweeted status)}
\usage{
parseTweets(tweets, simplify = FALSE, verbose = TRUE)
}
\arguments{
\item{tweets}{A character string naming the file where tweets are stored or the
name of the object in memory where the tweets were saved as strings.}

\item{simplify}{If \code{TRUE} it will return a data frame with only tweet and user
fields (i.e., no geographic information or url entities).}

\item{verbose}{logical, default is \code{TRUE}, which will print in the console
the number of tweets that have been parsed.}
}
\description{
Modified version of parseTweets that will return retweet
count for tweet (not retweeted status)
}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}
}

